+++ automad/src/server/Controllers/API/FileCollectionController.php
@@ -43,6 +43,7 @@ use Automad\Core\Messenger;
 use Automad\Core\Request;
 use Automad\Core\Text;
 use Automad\Models\FileCollection;
+use Automad\System\DiskUsage;
 
 defined('AUTOMAD') or die('Direct access not permitted!');
 
@@ -105,7 +106,7 @@ class FileCollectionController {
 	public static function upload(): Response {
 		$Response = new Response();
 
-		if (FileSystem::diskQuotaExceeded()) {
+		if (DiskUsage::quotaExceeded()) {
 			return $Response->setError(Text::get('diskQuotaExceeded'))->setCode(403);
 		}